Review a vacancy

+++ src/SFA.Apprenticeships.Infrastructure.Repositories.Sql.Tests/Schemas/Vacancy/VacancyRepositoryTests.cs
@@ -56,7 +56,7 @@
 
             writeRepository.Save(vacancy);
 
-            var entity = readRepository.GetByReferenceNumber((int) vacancy.VacancyReferenceNumber);
+            var entity = readRepository.GetByReferenceNumber(vacancy.VacancyReferenceNumber);
 
             entity.ShouldBeEquivalentTo(vacancy, options =>
                 ForShallowSave(options)
